Add V8 to recursedeps
It turns out that v8's DEPS hooks do work with recursedeps, as long as v8 is in the top
level directory of the solution (e.g. v8 instead of emscripten-releases/v8).
So this change is sufficient to build v8 in the recipe without any weird
gclient solutions.

Bug: v8:8921
Change-Id: Ifab9ecb17345f7373c7b9eff8397d864ff67f4a9
Reviewed-on: https://chromium-review.googlesource.com/c/emscripten-releases/+/1592100
Reviewed-by: Sergiy Belozorov <sergiyb@chromium.org>
Commit-Queue: Derek Schuff <dschuff@chromium.org>